requirements.txt
django
django-ninja
pytest
pytest-django
psycopg[binary]
python-dotenv
redis